requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
markitdown[pdf]
pypdfium2==4.30.0
pypdf==5.1.0
pillow==11.3.0
openai==1.59.5
easyocr==1.7.2
